src/Core/UseCases/SendInsertSucceededUseCase.cs
using System;
using StorageSimulator.Core.Interfaces;
using StorageSimulator.Core.Model;
using StorageSimulator.Core.Types;

namespace StorageSimulator.Core.UseCases
{
    public class SendInsertSucceededUseCase : ISendInsertSucceededUseCase
    {
        private readonly ISendResponseUseCase _sendResponseUseCase;

        public SendInsertSucceededUseCase(ISendResponseUseCase sendResponseUseCase)
        {
            _sendResponseUseCase = sendResponseUseCase;
        }

        public void Execute(MovementRequest request)
        {
            var response = new MovementResponse
            {
                Info = $"Insert: {request.Info}", Quantity = request.Quantity, Source = request.Source,
                Status = AutomationStatus.InsertionSucceeded, Target = request.Target, TargetCompartment = request.TargetCompartment,
                SourceCompartment = request.SourceCompartment, Ticket = request.Ticket, Timestamp = DateTime.UtcNow
            };
            _sendResponseUseCase.Execute(response);
        }
    }
}